+++ res/xml/development_settings.xml
@@ -635,10 +635,6 @@
             android:summary="@string/hwui_force_dark_summary" />
 
         <SwitchPreferenceCompat
-            android:key="enable_blurs_on_windows"
-            android:title="@string/window_blurs"/>
-
-        <SwitchPreferenceCompat
             android:key="force_msaa"
             android:title="@string/force_msaa"
             android:summary="@string/force_msaa_summary" />
