+++ res/xml/display_settings.xml
@@ -236,6 +236,10 @@
             android:title="@string/lift_to_wake_title"/>
 
         <SwitchPreferenceCompat
+            android:key="enable_blurs_on_windows"
+            android:title="@string/window_blurs"/>
+
+        <SwitchPreferenceCompat
             android:key="tap_to_wake"
             android:title="@string/tap_to_wake"
             android:summary="@string/tap_to_wake_summary"/>
